09-aws/requirements.txt
pandas
ipython
boto3 
python-dotenv